// packages/study-echarts/src/echarts/util/types.ts
export type ColorString = string;
export type ZRColor = ColorString | LinearGradientObject | RadialGradientObject | PatternObject;
// ...
export interface Payload extends PayloadItem {
  type: string;
  escapeConnect?: boolean;
  batch?: PayloadItem[];
}
// ...
export interface StageHandlerReset {
  (seriesModel: SeriesModel, ecModel: GlobalModel, api: ExtensionAPI, payload?: Payload):
    StageHandlerProgressExecutor | StageHandlerProgressExecutor[] | void
}
export interface StageHandlerOverallReset {
  (ecModel: GlobalModel, api: ExtensionAPI, payload?: Payload): void
}

export interface StageHandler {
  createOnAllSeries?: boolean;

  seriesType?: string;

  getTargetSeries?: (ecModel: GlobalModel, api: ExtensionAPI) => HashMap<SeriesModel>;

  performRawSeries?: boolean;

  plan?: StageHandlerPlan;

  overallReset?: StageHandlerOverallReset;

  reset?: StageHandlerReset;
}

export interface StageHandlerInternal extends StageHandler {
  uid: string;
  visualType?: 'layout' | 'visual';
  // modifyOutputEnd?: boolean;
  __prio: number;
  __raw: StageHandler | StageHandlerOverallReset;
  isVisual?: boolean; // PENDING: not used
  isLayout?: boolean; // PENDING: not used
}


export type StageHandlerProgressParams = TaskProgressParams;
export interface StageHandlerProgressExecutor {
  dataEach?: (data: SeriesData, idx: number) => void;
  progress?: (params: StageHandlerProgressParams, data: SeriesData) => void;
}
export type StageHandlerPlanReturn = TaskPlanCallbackReturn;
export interface StageHandlerPlan {
  (seriesModel: SeriesModel, ecModel: GlobalModel, api: ExtensionAPI, payload?: Payload):
    StageHandlerPlanReturn
}
// ...
export type DimensionName = string;
// ...
export type SeriesDataType = 'main' | 'node' | 'edge';
// ...
export type OptionDataItem =
  OptionDataValue
  | Dictionary<OptionDataValue>
  | OptionDataValue[]
  // FIXME: In some case (markpoint in geo (geo-map.html)), dataItem is {coord: [...]}
  | OptionDataItemObject<OptionDataValue>;
// Only for `SOURCE_FORMAT_KEYED_ORIGINAL`
export type OptionDataItemObject<T> = {
  id?: OptionId;
  name?: OptionName;
  groupId?: OptionId;
  value?: T[] | T;
  selected?: boolean;
};
// ...
export type OptionDataValue = string | number | Date;

export type OptionDataValueNumeric = number | '-';
export type OptionDataValueCategory = string;
export type OptionDataValueDate = Date | string | number;

// export type ModelOption = Dictionary<any> | any[] | string | number | boolean | ((...args: any) => any);
export type ModelOption = any;
export type ThemeOption = Dictionary<any>;

export type DisplayState = 'normal' | 'emphasis' | 'blur' | 'select';
// ...
export interface CallbackDataParams {
  // component main type
  componentType: string;
  // component sub type
  componentSubType: string;
  componentIndex: number;
  // series component sub type
  seriesType?: string;
  // series component index (the alias of `componentIndex` for series)
  seriesIndex?: number;
  seriesId?: string;
  seriesName?: string;
  name: string;
  dataIndex: number;
  data: OptionDataItem;
  dataType?: SeriesDataType;
  value: OptionDataItem | OptionDataValue;
  color?: ZRColor;
  borderColor?: string;
  dimensionNames?: DimensionName[];
  encode?: DimensionUserOuputEncode;
  marker?: TooltipMarker;
  status?: DisplayState;
  dimensionIndex?: number;
  percent?: number; // Only for chart like 'pie'

  // Param name list for mapping `a`, `b`, `c`, `d`, `e`
  $vars: string[];
}
// ...
export type SymbolSizeCallback<T> = (rawValue: any, params: T) => number | number[];
export type SymbolCallback<T> = (rawValue: any, params: T) => string;
export type SymbolRotateCallback<T> = (rawValue: any, params: T) => number;
export type SymbolOffsetCallback<T> = (rawValue: any, params: T) => string | number | (string | number)[];
/**
* Mixin of option set to control the element symbol.
* Include type of symbol, and size of symbol.
*/
export interface SymbolOptionMixin<T = never> {
  /**
   * type of symbol, like `cirlce`, `rect`, or custom path and image.
   */
  symbol?: string | (T extends never ? never : SymbolCallback<T>)
  /**
   * Size of symbol.
   */
  symbolSize?: number | number[] | (T extends never ? never : SymbolSizeCallback<T>)

  symbolRotate?: number | (T extends never ? never : SymbolRotateCallback<T>)

  symbolKeepAspect?: boolean

  symbolOffset?: string | number | (string | number)[] | (T extends never ? never : SymbolOffsetCallback<T>)
}
// ...
export interface DefaultStatesMixin {
  // FIXME
  emphasis?: any
  select?: any
  blur?: any
}

export type DefaultEmphasisFocus = 'none' | 'self' | 'series';

export interface DefaultStatesMixinEmphasis {
  /**
   * self: Focus self and blur all others.
   * series: Focus series and blur all other series.
   */
  focus?: DefaultEmphasisFocus
}

export interface StatesMixinBase {
  emphasis?: unknown
  select?: unknown
  blur?: unknown
}
// ...
export interface SeriesOption<
  StateOption = unknown,
  StatesMixin extends StatesMixinBase = DefaultStatesMixin
  > extends